WIP Resumable upload
Clean up handler factories
